refactor(validation & load): validation and loading utils to be generic
Move from the overly specific (and inaccurate, now that we're parsing
timeseries) "scenarios" to generic "data" for validation and loading,
including passing generic type

Also rename functions ass appropriate

Instead, when trying to load files that match a particular schema, pass
in that schema
