+++ src/data/scenariosData.ts
@@ -1,9 +1,7 @@
 import { Scenario } from "../types";
-import { FileEntry } from "../utils/validateScenarios";
-import {
-  assembleScenarios,
-  decideIncludeInvalid,
-} from "../utils/loadScenarios";
+import { FileEntry } from "../utils/validateData";
+import { assembleData, decideIncludeInvalid } from "../utils/loadData";
+import pathwayMetadata from "../schema/pathwayMetadata.v1.json" with { type: "json" };
 
 // 1) Grab every JSON file in this folder **and subfolders**
 //    `eager:true` = load at build time (no async), `import:'default'` = get the parsed JSON
@@ -21,6 +19,10 @@ const entries: FileEntry[] = Object.entries(modules)
   }))
   .sort((a, b) => a.name.localeCompare(b.name));
 
-export const scenariosData: Scenario[] = assembleScenarios(entries, {
-  includeInvalid: decideIncludeInvalid(),
-});
+export const scenariosData: Scenario[] = assembleData(
+  entries,
+  pathwayMetadata,
+  {
+    includeInvalid: decideIncludeInvalid(),
+  },
+);
